+++ politics_app/src/hooks/useScrollDetection.tsx
@@ -1,0 +1,19 @@
+import { useState, useEffect } from "react";
+
+export const useScrollDetection = (threshold: number = 10) => {
+  const [isScrolled, setIsScrolled] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const scrolled = window.scrollY > threshold;
+      if (scrolled !== isScrolled) {
+        setIsScrolled(scrolled);
+      }
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, [isScrolled, threshold]);
+
+  return isScrolled;
+};
